@@ -1,10 +1,7 @@
 import './Footer.css'
-import React from 'react'
+import React from "react"
 
 export default props =>
     <footer className="footer">
-        <span>
-            Desenvolvido com <i className="fa fa-heart text-danger"></i> por
-            <strong> Cod<span className="text-danger">3</span>r</strong>
-        </span>
+        Rodapé
     </footer>
