@@ -1,11 +1,7 @@
 import './Logo.css'
-import logo from '../../assets/imgs/logo.png'
-import React from 'react'
-import { Link } from 'react-router-dom'
+import React from "react";
 
 export default props =>
     <aside className="logo">
-        <Link to="/" className="logo">
-            <img src={logo} alt="logo" />
-        </Link>
+        Logo
     </aside>
