@@ -1,13 +1,11 @@
 import './Main.css'
-import React from 'react'
-import Header from './Header'
+import React from "react";
+import Header from "./Header";
 
 export default props =>
     <React.Fragment>
-        <Header {...props} />
-        <main className="content container-fluid">
-            <div className="p-3 mt-3">
-                {props.children}
-            </div>
+        <Header/>
+        <main className="content">
+            Conteúdo
         </main>
     </React.Fragment>
