@@ -1,20 +1,15 @@
-import 'bootstrap/dist/css/bootstrap.min.css'
-import 'font-awesome/css/font-awesome.min.css'
 import './App.css'
-import React from 'react'
-import { BrowserRouter } from 'react-router-dom'
+import React from "react"
 
+import Main from "../components/template/Main"
 import Logo from '../components/template/Logo'
-import Nav from '../components/template/Nav'
-import Routes from './Routes'
-import Footer from '../components/template/Footer'
+import Nav from "../components/template/Nav"
+import Footer from "../components/template/Footer"
 
 export default props =>
-    <BrowserRouter>
-        <div className="app">
-            <Logo />
-            <Nav />
-            <Routes />
-            <Footer />
-        </div>
-    </BrowserRouter>
+    <div className="app">
+        <Logo/>
+        <Nav/>
+        <Main/>
+        <Footer/>
+    </div>
